+++ .idea/encodings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/src/puyopuyo/Panel/map/roundthread/RoundThread.java" charset="x-windows-949" />
     <file url="file://$PROJECT_DIR$/src/puyopuyo/ScorePanel.java" charset="x-windows-949" />
-    <file url="file://$PROJECT_DIR$/src/puyopuyo/map/roundthread/RoundThread.java" charset="x-windows-949" />
   </component>
 </project>
