--- src/puyopuyo/Constants.java
@@ -1,7 +1,0 @@
-package puyopuyo;
-
-public final class Constants {
-    public static final int[] COMBO_BONUS = {0,0,8,16,32,64,96,128,160,192,224,256,288,320,352,384,416,448,480,512};
-    public static final int[] CONNECT_BONUS = {0,0,0,0,0,2,3,4,5,6,7,10};
-    public static final int[] COLOR_BONUS = {0,3,6,12,24};
-}
