+++ src/puyopuyo/Panel/map/MapPanel.java
@@ -1,13 +1,20 @@
-package puyopuyo.map;
+package puyopuyo.Panel.map;
 
 import puyopuyo.frame.Frame;
-import puyopuyo.frame.PanelState;
-import puyopuyo.game.GameThread;
-import puyopuyo.puyo.GameImageIcon;
+import puyopuyo.Panel.PanelState;
+import puyopuyo.Panel.map.game.GameThread;
+import puyopuyo.resource.GameImageIcon;
 
 import javax.swing.*;
 import java.awt.*;
 
+/**
+ * 게임 전체 화면을 관리하는 패널이다.<br>
+ *
+ * 시각적인 책임을 가지며, MVC 패턴의 View의 역할을 한다.
+ *
+ * 게임 창은 하나만 구현되므로, 싱글톤 패턴을 이용하여 개발했다.
+ */
 public class MapPanel extends JPanel implements PanelState {
     private static MapPanel instance;
 
@@ -23,12 +30,26 @@ public class MapPanel extends JPanel implements PanelState {
         process();
     }
 
+    /**
+     * 패널의 백그라운드 이미지를 그리기 위한 함수 <br>
+     * 전체 화면 크기로 이미지가 배치 된다.
+     */
     @Override
     public void paintComponent(Graphics graphics) {
         super.paintComponent(graphics);
         graphics.drawImage(GameImageIcon.gamePanelIcon.getImage(), 0, 0, this.getWidth(), this.getHeight(), this);
     }
 
+    /**
+     * 화면 설정을 하는 함수이다. <br>
+     *
+     * 1. 플레이어가 진행할 GameGround, ScorePanel을 생성한다. <br>
+     * 2. SubPanel 위치: <br>
+     *   __1) GameGround1P: 50, 60 <br>
+     *   __2) GameGround2P: 830, 60 <br>
+     *   __3) ScorePanel: 490, 60 <br>
+     * 3. 해당 패널에 ControlPuyoKeyListener 연결한다. <br>
+     */
     @Override
     public void setUi() {
         this.setLayout(null);
@@ -75,13 +96,18 @@ public class MapPanel extends JPanel implements PanelState {
         frame.remove(me);
     }
 
+    /**
+     * 해당 패널에서 키보드를 주관할 수 있게 만드는 함수이다.
+     *
+     */
     public void addKeyListener() {
         /*
          * 해당 패널의 키보드 값을 받도록 설정.
          * 보통은 따로하나, 동시에 1p, 2p의 키보드 값을 받게하기 위함
          */
-        addKeyListener(new ControlPuyoKeyListener()); //TODO: 수정할 것
+        //TODO: 에러발생 위치:: 수정할 것
+        addKeyListener(new ControlPuyoKeyListener());
         setFocusable(true);
-        //mapPanel.requestFocus();
+        requestFocus();
     }
 }
