+++ src/puyopuyo/Panel/map/subpanel/ground/round/RoundRepository.java
@@ -1,4 +1,4 @@
-package puyopuyo.round;
+package puyopuyo.Panel.map.subpanel.ground.round;
 
 public class RoundRepository {
     private final int iAm;
