+++ src/puyopuyo/PuyoPuyo.java
@@ -1,13 +1,13 @@
 package puyopuyo;
 
 import puyopuyo.frame.Frame;
-import puyopuyo.menu.MenuPanel;
+import puyopuyo.Panel.menu.StartPanel;
 
 public class PuyoPuyo {
 	public static void main(String[] args) {
 		var frame = Frame.getInstance();
 
-		var menuPanel = MenuPanel.getInstance();
-		frame.changePanel(menuPanel);
+		var startPanel = StartPanel.getInstance();
+		frame.changePanel(startPanel);
 	}
 }
