--- src/puyopuyo/frame/PanelState.java
@@ -1,8 +1,0 @@
-package puyopuyo.frame;
-
-public interface PanelState {
-    void setUi();
-    void process();
-    void open();
-    void close();
-}
